# oficina/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Operare API
  description: |
    API do sistema Operare — gestão para oficinas mecânicas automotivas.
    Multi-tenant com Row-Level Security. Todas as rotas requerem autenticação via NextAuth JWT (exceto /api/public/*).
  version: "1.0.0"
  contact:
    name: DF Developer
    email: contato@operare.tech
servers:
  - url: http://localhost:3000/api
    description: Desenvolvimento local
  - url: https://operare.tech/api
    description: Produção
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT via NextAuth (cookie ou header)
  schemas:
...
    Vehicle:
      type: object
      properties:
        id: { type: string, format: cuid }
        plate: { type: string, example: "ABC1D23" }
        brand: { type: string, example: "Fiat" }
        model: { type: string, example: "Uno" }
        year: { type: integer, example: 2020 }
        color: { type: string, nullable: true }
        mileage: { type: integer, example: 45000 }
        clientId: { type: string }
        tenantId: { type: string }
    OrderSummary:
      type: object
      properties:
        id: { type: string }
        number: { type: integer, example: 42 }
        status: { type: string, enum: [OPEN, IN_PROGRESS, WAITING_PART, WAITING_APPROVAL, COMPLETED, DELIVERED, CANCELLED] }
        totalAmount: { type: number, format: float, example: 1250.50 }
        createdAt: { type: string, format: date-time }
        client: { type: object, properties: { name: { type: string } } }
        vehicle: { type: object, properties: { plate: { type: string }, model: { type: string } } }
    Complaint:
      type: object
      properties:
        description: { type: string, example: "Barulho na suspensão" }
        services:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              price: { type: number }
              mechanicId: { type: string, nullable: true }
              serviceId: { type: string, nullable: true }
        parts:
          type: array
          items:
            type: object
            properties:
              description: { type: string }
              quantity: { type: integer }
              unitPrice: { type: number }
              stockItemId: { type: string, nullable: true }
...
    PaginatedOrders:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/OrderSummary" }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
security:
  - session: []
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: Listar veículos
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Vehicle" }
    post:
      tags: [Vehicles]
      summary: Criar veículo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plate, brand, model, year, clientId]
              properties:
                plate: { type: string }
                brand: { type: string }
                model: { type: string }
                year: { type: integer }
                color: { type: string }
                clientId: { type: string }
      responses:
        "201": { description: Veículo criado }
        "409": { description: Placa já cadastrada neste tenant }

  # ─── ORDERS ─────────────────────────────────────────────────────
  /orders:
    get:
      tags: [Orders]
      summary: Listar ordens de serviço (paginado)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [OPEN, IN_PROGRESS, WAITING_PART, WAITING_APPROVAL, COMPLETED, DELIVERED, CANCELLED] }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
        - name: mechanicId
          in: query
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, default: 15 }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedOrders" }
    post:
      tags: [Orders]
      summary: Criar ordem de serviço
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId, vehicleId]
              properties:
                clientId: { type: string }
                vehicleId: { type: string }
                mileage: { type: integer }
                notes: { type: string }
                complaints:
                  type: array
                  items: { $ref: "#/components/schemas/Complaint" }
      responses:
        "201":
          description: OS criada (com possíveis stockWarnings)
        "400": { description: Validação falhou }
...
  /orders/export:
    get:
      tags: [Orders]
      summary: Exportar OSs em CSV
      parameters:
        - name: ids
          in: query
          schema: { type: array, items: { type: string } }
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Arquivo CSV
          content:
            text/csv: {}
...
  /reports:
    get:
      tags: [Reports]
      summary: Relatório financeiro (admin)
      parameters:
        - name: startDate
          in: query
          schema: { type: string, format: date }
        - name: endDate
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Dados financeiros agregados
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOrders: { type: integer }
                  totalRevenue: { type: number }
                  avgTicket: { type: number }
                  partsCost: { type: number }
                  grossProfit: { type: number }
                  cancelledCount: { type: integer }
                  completedCount: { type: integer }
                  byStatus: { type: object }
                  monthly: { type: array, items: { type: object, properties: { month: { type: string }, revenue: { type: number }, count: { type: integer } } } }
                  profitByOrder: { type: array }
        "403": { description: Apenas admin }
...
  /users:
    get:
      tags: [Users]
      summary: Listar usuários do tenant
      responses:
        "200":
          description: Lista de usuários
    post:
      tags: [Users]
      summary: Criar usuário (admin)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, email, role, password]
              properties:
                name: { type: string, minLength: 3 }
                email: { type: string, format: email }
                role: { type: string, enum: [ADMIN, MECHANIC, ATTENDANT] }
                password: { type: string, minLength: 8 }
                commissionRate: { type: number }
      responses:
        "201": { description: Usuário criado }
        "409": { description: E-mail já em uso }
...