# oficina/src/app/api/vehicles/route.ts
import { NextRequest, NextResponse } from "next/server";
import { createContainer } from "@/infrastructure/container";
import { CreateVehicle } from "@/application/use-cases/vehicles/CreateVehicle";
import { handleError } from "@/lib/api-handler";
import { requireAuth } from "@/lib/auth";

export async function GET(request: NextRequest) {
  try {
    const session = await requireAuth();
    const tenantId = session.user.tenantId;
    const container = createContainer(tenantId);

    const search = request.nextUrl.searchParams.get("search") || "";

    const vehicles = search
      ? await container.vehicleRepository.search(search, tenantId)
      : await container.vehicleRepository.findAll(tenantId);

    return NextResponse.json(vehicles);
  } catch (error) {
    return handleError(error);
  }
}

export async function POST(request: NextRequest) {
  try {
    const session = await requireAuth();
    const tenantId = session.user.tenantId;
    const container = createContainer(tenantId);

    const body = await request.json();
    const useCase = new CreateVehicle(container.vehicleRepository);
    const vehicle = await useCase.execute(body, tenantId);
    return NextResponse.json(vehicle, { status: 201 });
  } catch (error) {
    return handleError(error);
  }
}